chore: Move artwork code for "Now Playing" screen to its own file
